Feat: 添加KDJ数据计算逻辑

+++ KDS_Phone/Views/JT_KLineView.h
@@ -27,29 +27,29 @@
 /**
  顶部5日均线、10日均线显示区域的高度
  */
-@property (nonatomic ,assign ) CGFloat MALineHeight;
+@property (nonatomic ,assign ) float MALineHeight;
 
 
 /**
  右侧选择器的宽度
  */
-@property (nonatomic ,assign ) CGFloat rightSelecterWidth;
+@property (nonatomic ,assign ) float rightSelecterWidth;
 
 /**
  k 线蜡烛线视图高度
  */
-@property (nonatomic ,assign) CGFloat klineChartViewHeight;
+@property (nonatomic ,assign) float klineChartViewHeight;
 
 /**
  中间时间区域的高度
  */
-@property (nonatomic ,assign) CGFloat timeViewHeight;
+@property (nonatomic ,assign) float timeViewHeight;
 
 
 /**
  成交量底部的Margin
  */
-@property (nonatomic ,assign) CGFloat bottomMargin;
+@property (nonatomic ,assign) float bottomMargin;
 
 
 /**
@@ -60,13 +60,13 @@
 /**
  下部指标选择区高度
  */
-@property (nonatomic ,assign) CGFloat indicatorViewHeight;
+@property (nonatomic ,assign) float indicatorViewHeight;
 
 @property(nonatomic, copy) NSArray<JT_KLineModel *> *kLineModels;
 // k 线是否需要拖动和缩放
 @property (nonatomic ,assign) BOOL needZoomAndScroll;
 //可以设置蜡烛线区域top And Bottom Margin
-@property (nonatomic ,assign ) CGFloat KlineChartTopMargin;
+@property (nonatomic ,assign ) float KlineChartTopMargin;
 - (void)reDrawAllView;
 
 @end
